pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel"
version = "0.1.0"
description = "Autonomous Supply Chain Intelligence Agent"
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "prefect",
    "groq",
    "feedparser",
    "trafilatura",
    "requests",
    "python-dotenv",
    "streamlit",
    "pydantic",
]

[tool.setuptools.packages.find]
where = ["src"]